% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.r
\name{addSourcesToDictionary}
\alias{addSourcesToDictionary}
\title{Add data sources information to the dictionary
from the data/sources.csv file.
See generateSourcesSpreadsheet() for details about creating this file. 
This information is used when decide which data values
to remove for participants who have withdrawn consent.}
\usage{
addSourcesToDictionary(
  dictionary,
  sourcesFile = system.file("data/extdata", "sources.csv", package = "alspac")
)
}
\arguments{
\item{dictionary}{The name of an existing dictionary or the dictionary itself.}
}
\description{
Add data sources information to the dictionary
from the data/sources.csv file.
See generateSourcesSpreadsheet() for details about creating this file. 
This information is used when decide which data values
to remove for participants who have withdrawn consent.
}
