% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alspac-package.R
\docType{package}
\name{alspac-package}
\alias{alspac}
\alias{alspac-package}
\title{alspac: Data dictionary for ALSPAC}
\description{
Functions to search and extract variables based on keywords from the ALSPAC data dictionary.
}
\author{
\strong{Maintainer}: Gibran Hemani \email{g.hemani@bristol.ac.uk}

Authors:
\itemize{
  \item Matthew Suderman \email{matt.suderman@bristol.ac.uk}
}

}
\keyword{internal}
