% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{current}
\alias{current}
\title{Current data}
\format{
## `current`
A data frame with 87,248 rows and 19 columns:
\describe{
  \item{obj}{Stata dataset name}
  \item{name}{variable name}
  \item{lab}{}
  \item{type}{}
  \item{counts}{}
  \item{cat1}{}
  \item{cat2}{}
  \item{cat3}{}
  \item{cat4}{}
  \item{path}{}
  \item{cat5}{}
  \item{mother}{}
  \item{mother_clinic}{}
  \item{mother_quest}{}
  \item{partner_quest}{}
  \item{partner_clinic}{}
  \item{partner}{}
  \item{child_based}{}
  \item{child_completed}{}
}
}
\usage{
current
}
\description{
A list of current datasets.
}
\keyword{datasets}
