% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAlspac.R
\name{extractWebOutput}
\alias{extractWebOutput}
\title{Extract variables exported from the ALSPAC variable lookup web app}
\usage{
extractWebOutput(filename)
}
\arguments{
\item{filename}{Name of file exported from ALSPAC variable lookup web app}
}
\value{
Data frame
}
\description{
The variable lookup webapp allows you to browse the available
variables and export a list of selected variables.
This function will read that exported list and extract the individual
level data for each of the selected variables.
}
\details{
More generally, this function requires a file that has at least one 
column with the header 'Variable' followed by a list of variable names.

The R: drive must be mounted and its path set with the \code{setDataDir} function.
}
