% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAlspac.R
\name{filterVars}
\alias{filterVars}
\title{Filter duplicate variables from findVars}
\usage{
filterVars(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{findVars}()}.}

\item{...}{Filter terms.  The name corresponds to the variable
name for which to remove duplicates.  Each term is a named vector
whose names correspond to columns in `x`.
The values provide patterns for the given column to match.}
}
\value{
The subset of `x` that satisfies the supplied filters
or that were not provided a filter.
}
\description{
Filter duplicate variables from findVars
}
\details{
\code{\link{findVars}()} may identify multiple
variables with the same name.  This function can be used
to select among these duplicates.
}
\examples{
\dontrun{
varnames <- c("kz021","kz011b","ype9670", "c645a")
vars <- findVars(varnames)
vars <- subset(vars, subset=tolower(name) \%in\% varnames)
vars <- filterVars(vars, kz021=c(obj="^kz"), kz011b=c(obj="^cp", lab="Participant"), c645a=c(cat2="Quest")) 
}
}
