% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAlspac.R
\name{findVars}
\alias{findVars}
\title{Find variables}
\usage{
findVars(..., logic = "any", ignore.case = TRUE, perl = FALSE,
  fixed = FALSE, whole.word = FALSE, dictionary = "current")
}
\arguments{
\item{...}{Search terms}

\item{logic}{Conditions for the search strings, can be "all", "any", or "none". Set to "any" by default.}

\item{ignore.case}{Should search terms be case sensitive? Defaults to TRUE.}

\item{perl}{logical.  Should perl-compatible regexps be used? Defaults to FALSE.}

\item{fixed}{logical.  If 'TRUE', 'pattern' is a string to be matched as is.  Overrides all conflicting arguments. Defaults to FALSE.}

\item{whole.word}{If 'TRUE' search term "word" will be changed to "\\bword\\b" to only match whole words. Defaults to FALSE.}

\item{dictionary}{Data frame or name of a data dictionary. Defaults to "current", but "useful" is also an option - this contains the data dictionary for the R:/Data/Useful_data directory. New dictionaries can be created using the \code{\link{createDictionary}()} function.}
}
\value{
A data frame containing a list of the variables, the files they originate from, and some descripton about the files
}
\description{
Provide a list of search terms to find variables in the data dictionary.
}
\details{
The resulting data frame will have the following columns:
\itemize{
	\item{obj - The name of the data file}
	\item{name - The name of the variable}
	\item{type - The type of data for the variable}
	\item{lab - A description of the label}
	\item{code - The ALSPAC dictionary code for the variable}
	\item{counts - The number of non-NA values in the variable}
	\item{cat1-4 - These columns correspond to the folder names that the objects were found in}
}
}
\examples{
\dontrun{
# Find variables with BMI or height in the description (this will return a lot of results!)
bmi_variables <- findVars("bmi", "height", logic="any", ignore.case=TRUE)
}
}
