% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.r
\name{generateSourcesSpreadsheet}
\alias{generateSourcesSpreadsheet}
\title{This function was used to initially create the data/sources.csv
spreadsheet which provides the source of data for each data file ('obj')
in the dictionary. This information is then used to determine 
which bits of data to remove to satisfy exclusion lists. 
Sources include mother_clinic, mother_quest, partner, partner_clinic, partner_quest,
child_based and child_completed.
Sources can for the most part be determined automatically from the 'path'
information provided for each variable in the dictionary.
However, for variables in "Useful_data", the source isn't always clear 
so needs to be identified manually.}
\usage{
generateSourcesSpreadsheet()
}
\description{
sources <- generateSourcesSpreadsheet()
write.csv(sources, file="data/sources.csv", row.names=F)
}
