% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadir.R
\name{getDefaultDataDir}
\alias{getDefaultDataDir}
\title{Guess the default data directory}
\usage{
getDefaultDataDir()
}
\description{
The R drive will be mounted in different paths for different systems.
This function guesses the path to be used as default in setDataDir
}
\examples{
\dontrun{

}
}
