% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.r
\name{removeExclusions}
\alias{removeExclusions}
\title{Remove data for participants who have withdrawn consent.}
\usage{
removeExclusions(x)
}
\arguments{
\item{x}{Data frame output from \code{\link{extractVars}()}.}
}
\value{
The input data frame but with appropriate values set to missing
with additional variables ("woc_*") identifying participants
who have withdrawn consent.
}
\description{
The exclusion lists for mothers and children are stored in .do
files in the R: drive. This function obtains ALNs to exclude
from these files and then sets the variable
values to missing for the appropriate participants
and adds indicator variables for these participants ("woc_*").
}
