% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadir.R
\name{setDataDir}
\alias{setDataDir}
\title{Set the data directory}
\usage{
setDataDir(datadir = getDefaultDataDir())
}
\arguments{
\item{datadir}{The directory where the ALSPAC data can be found}
}
\value{
Null. Assigns the option alspac_data_dir
}
\description{
This function is automatically called upon loading the package through `library(alspac)`
It creates a global option called `alspac_data_dir` which is used by the \code{extractVars}
function to locate the alspac data files.
This function guesses the path to be used as default in setDataDir. The defaults are:
\itemize{
\item{Windows: R:/Data/Useful_data/current_R/}
\item{Mac: /Volumes/data/Useful_data/current_R/}
\item{Linux: ~/.gvfs/data/Useful_data/current_R/}
}
}
\examples{
\dontrun{
setDataDir() # This sets the path based on the operating system's default
setDataDir("/some/other/path/") # This is how to supply a path manually
}
}
