% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclusions.r
\name{readExclusions}
\alias{readExclusions}
\title{Get list of ALNs to exclude}
\usage{
readExclusions()
}
\value{
List of ALNs for each .do file.
}
\description{
The exclusion lists for mothers and children are stored in .do
files in the R: drive. This function reads all of these .do files
and then parses out the ALNS for withdrawn consent.
}
